chore: reorganize key-vault code
