+++ key-vault/Cargo.toml
@@ -11,7 +11,7 @@ license = "MIT"
 crate-type = ["cdylib"]
 
 [dependencies]
-fips205 = { version = "0.4.1" }
+fips205 = { version = "0.4.1", default-features = false }
 aes-gcm = { version = "0.10.3", features = ["aes", "zeroize"] }
 hex = "0.4.3"
 scrypt = { version = "0.11.0", default-features = false }
